# PERF: Optimize point transformations