--- a/.pylintrc
+++ b/.pylintrc
@@ -23,4 +23,4 @@ disable=cyclic-import,
 
 
 [FORMAT]
-max-module-lines=1350
+max-module-lines=1500