Related: #1202
```python
In [1]: from pyproj import Transformer

In [2]: transformer = Transformer.from_crs(4326, 3857, always_xy=True)

In [3]: %timeit transformer.transform(1, 2)
4.69 µs ± 37.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [4]: %timeit transformer.transform_point(1, 2)
2.08 µs ± 20.2 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```
Version in #1203:
```python
In [5]: %timeit transformer.transform_point(1, 2)
2.99 µs ± 132 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```