@snowman2 2022-12-20
With #1208 

```
In [3]: transformer = Transformer.from_crs(4326, 3857, always_xy=True)

In [4]: %timeit transformer.transform(1, 2)
1.44 µs ± 38.6 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [5]: %timeit transformer.transform(1, 2)
1.42 µs ± 14.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)

In [6]: %timeit transformer.transform(1, 2)
1.44 µs ± 19.1 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```