@snowman2 2022-12-20
> I do like this better than adding a new method, and theoretically downstream users should be able to get immediate benefits without needing to change the call sites.

I agree